Small improvements

+++ src/MakingDotNETApplicationsFaster/Infrastructure/DemoRunner.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -6,14 +5,14 @@ namespace MakingDotNETApplicationsFaster.Infrastructure
 {
     static class DemoRunner
     {
-        static Dictionary<Int16, IRunner> Runners { get; set; }
+        static Dictionary<int, IRunner> Runners { get; set; }
 
         static DemoRunner()
         {
-            Runners = new Dictionary<Int16, IRunner>();
+            Runners = new Dictionary<int, IRunner>();
         }
 
-        public static bool TryAddRunner(Int16 id, IRunner runner)
+        public static bool TryAddRunner(int id, IRunner runner)
         {
             if (Runners.ContainsKey(id))
             {
@@ -23,12 +22,12 @@ namespace MakingDotNETApplicationsFaster.Infrastructure
             return true;
         }
 
-        public static bool TryAddRunners(Dictionary<Int16, IRunner> runnersMap)
+        public static bool TryAddRunners(Dictionary<int, IRunner> runnersMap)
         {
             return runnersMap.Select(runner => TryAddRunner(runner.Key, runner.Value)).All(isSuccess => isSuccess);
         }
 
-        public static bool Run(Int16 runnerId)
+        public static bool Run(int runnerId)
         {
             IRunner runner;
             if (Runners.TryGetValue(runnerId, out runner))
